Sync with HabitRPG:develop

+++ Habitica/src/main/java/com/habitrpg/android/habitica/ui/activities/TaskFormActivity.kt
@@ -1,54 +1,63 @@
 package com.habitrpg.android.habitica.ui.activities
 
+import android.app.Activity
 import android.content.Context
+import android.content.Intent
+import android.content.res.ColorStateList
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
+import android.os.Handler
 import android.view.Menu
+import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
+import android.view.inputmethod.InputMethodManager
 import android.widget.CheckBox
 import android.widget.EditText
 import android.widget.LinearLayout
 import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.widget.AppCompatCheckBox
 import androidx.appcompat.widget.Toolbar
 import androidx.core.content.ContextCompat
+import androidx.core.view.children
+import androidx.core.view.forEachIndexed
+import androidx.core.widget.NestedScrollView
 import com.habitrpg.android.habitica.R
 import com.habitrpg.android.habitica.components.AppComponent
 import com.habitrpg.android.habitica.data.TagRepository
 import com.habitrpg.android.habitica.data.TaskRepository
 import com.habitrpg.android.habitica.data.UserRepository
-import com.habitrpg.android.habitica.extensions.notNull
+import com.habitrpg.android.habitica.extensions.OnChangeTextWatcher
+import com.habitrpg.android.habitica.extensions.dpToPx
 import com.habitrpg.android.habitica.helpers.RxErrorHandler
+import com.habitrpg.android.habitica.helpers.TaskAlarmManager
 import com.habitrpg.android.habitica.models.Tag
 import com.habitrpg.android.habitica.models.tasks.HabitResetOption
 import com.habitrpg.android.habitica.models.tasks.Task
 import com.habitrpg.android.habitica.models.user.Stats
 import com.habitrpg.android.habitica.ui.helpers.bindView
-import io.reactivex.functions.Consumer
-import javax.inject.Inject
-import android.content.res.ColorStateList
-import android.graphics.drawable.ColorDrawable
-import android.view.MenuItem
-import android.view.inputmethod.InputMethodManager
-import androidx.appcompat.widget.AppCompatCheckBox
-import androidx.core.view.children
-import androidx.core.view.forEach
-import androidx.core.view.forEachIndexed
-import com.habitrpg.android.habitica.extensions.dpToPx
 import com.habitrpg.android.habitica.ui.views.tasks.form.*
+import io.reactivex.functions.Consumer
 import io.realm.RealmList
 import java.util.*
+import javax.inject.Inject
 
 
 class TaskFormActivity : BaseActivity() {
 
+    private var isSaving: Boolean = false
     @Inject
     lateinit var userRepository: UserRepository
     @Inject
     lateinit var taskRepository: TaskRepository
     @Inject
     lateinit var tagRepository: TagRepository
+    @Inject
+    lateinit var taskAlarmManager: TaskAlarmManager
 
     private val toolbar: Toolbar by bindView(R.id.toolbar)
+    private val scrollView: NestedScrollView by bindView(R.id.scroll_view)
     private val upperTextWrapper: LinearLayout by bindView(R.id.upper_text_wrapper)
     private val textEditText: EditText by bindView(R.id.text_edit_text)
     private val notesEditText: EditText by bindView(R.id.notes_edit_text)
@@ -78,9 +87,11 @@ class TaskFormActivity : BaseActivity() {
     private val rewardValueTitleView: TextView by bindView(R.id.reward_value_title)
     private val rewardValueFormView: RewardValueFormView by bindView(R.id.reward_value)
 
+    private val tagsTitleView: TextView by bindView(R.id.tags_title)
     private val tagsWrapper: LinearLayout by bindView(R.id.tags_wrapper)
 
     private var isCreating = true
+    private var isChallengeTask = false
     private var usesTaskAttributeStats = false
     private var task: Task? = null
     private var taskType: String = ""
@@ -90,6 +101,8 @@ class TaskFormActivity : BaseActivity() {
         field = value
         setSelectedAttribute(value)
     }
+
+    private var canSave: Boolean = false
 
     private var tintColor: Int = 0
     set(value) {
@@ -120,16 +133,10 @@ class TaskFormActivity : BaseActivity() {
 
         val bundle = intent.extras ?: return
 
-        taskType = bundle.getString(OldTaskFormActivity.TASK_TYPE_KEY) ?: Task.TYPE_HABIT
-        val taskId = bundle.getString(OldTaskFormActivity.TASK_ID_KEY)
-        if (taskId != null) {
-            isCreating = false
-            compositeSubscription.add(taskRepository.getUnmanagedTask(taskId).firstElement().subscribe(Consumer {
-                task = it
-                //tintColor = ContextCompat.getColor(this, it.mediumTaskColor)
-                fillForm(it)
-            }, RxErrorHandler.handleEmptyError()))
-        }
+        isChallengeTask = bundle.getBoolean(IS_CHALLENGE_TASK, false)
+
+        taskType = bundle.getString(TASK_TYPE_KEY) ?: Task.TYPE_HABIT
+        val taskId = bundle.getString(TASK_ID_KEY)
 
         compositeSubscription.add(tagRepository.getTags()
                 .map { tagRepository.getUnmanagedCopy(it) }
@@ -142,11 +149,30 @@ class TaskFormActivity : BaseActivity() {
             configureForm()
         }, RxErrorHandler.handleEmptyError()))
 
+
+        textEditText.addTextChangedListener(OnChangeTextWatcher { _, _, _, _ ->
+            checkCanSave()
+        })
         statStrengthButton.setOnClickListener { selectedStat = Stats.STRENGTH }
         statIntelligenceButton.setOnClickListener { selectedStat = Stats.INTELLIGENCE }
         statConstitutionButton.setOnClickListener { selectedStat = Stats.CONSTITUTION }
         statPerceptionButton.setOnClickListener { selectedStat = Stats.PERCEPTION }
+        scrollView.setOnScrollChangeListener { _: NestedScrollView?, _: Int, _: Int, _: Int, _: Int ->
+            dismissKeyboard()
+        }
 
+        if (taskId != null) {
+            isCreating = false
+            compositeSubscription.add(taskRepository.getUnmanagedTask(taskId).firstElement().subscribe(Consumer {
+                task = it
+                //tintColor = ContextCompat.getColor(this, it.mediumTaskColor)
+                fillForm(it)
+            }, RxErrorHandler.handleEmptyError()))
+        } else if (bundle.containsKey(PARCELABLE_TASK)) {
+            isCreating = false
+            task = bundle.getParcelable(PARCELABLE_TASK)
+            task?.let { fillForm(it) }
+        }
         configureForm()
     }
 
@@ -156,16 +182,24 @@ class TaskFormActivity : BaseActivity() {
         } else {
             menuInflater.inflate(R.menu.menu_task_edit, menu)
         }
+        menu.findItem(R.id.action_save).isEnabled = canSave
         return true
     }
 
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
         when (item?.itemId) {
-            R.id.action_create -> saveTask()
-            R.id.action_save_changes -> saveTask()
+            R.id.action_save -> saveTask()
             R.id.action_delete -> deleteTask()
         }
         return super.onOptionsItemSelected(item)
+    }
+
+    private fun checkCanSave() {
+        val newCanSave = textEditText.text.isNotBlank()
+        if (newCanSave != canSave) {
+            invalidateOptionsMenu()
+        }
+        canSave = newCanSave
     }
 
     private fun configureForm() {
@@ -186,11 +220,11 @@ class TaskFormActivity : BaseActivity() {
 
         val todoDailyViewsVisibility = if (taskType == Task.TYPE_DAILY || taskType == Task.TYPE_TODO) View.VISIBLE else View.GONE
 
-        checklistTitleView.visibility = todoDailyViewsVisibility
-        checklistContainer.visibility = todoDailyViewsVisibility
+        checklistTitleView.visibility = if (isChallengeTask) View.GONE else todoDailyViewsVisibility
+        checklistContainer.visibility = if (isChallengeTask) View.GONE else todoDailyViewsVisibility
 
-        remindersTitleView.visibility = todoDailyViewsVisibility
-        remindersContainer.visibility = todoDailyViewsVisibility
+        remindersTitleView.visibility = if (isChallengeTask) View.GONE else todoDailyViewsVisibility
+        remindersContainer.visibility = if (isChallengeTask) View.GONE else todoDailyViewsVisibility
 
         taskSchedulingTitleView.visibility = todoDailyViewsVisibility
         taskSchedulingControls.visibility = todoDailyViewsVisibility
@@ -203,6 +237,9 @@ class TaskFormActivity : BaseActivity() {
         val rewardViewsVisibility = if (taskType == Task.TYPE_REWARD) View.VISIBLE else View.GONE
         rewardValueTitleView.visibility = rewardViewsVisibility
         rewardValueFormView.visibility = rewardViewsVisibility
+
+        tagsTitleView.visibility = if (isChallengeTask) View.GONE else View.VISIBLE
+        tagsWrapper.visibility = if (isChallengeTask) View.GONE else View.VISIBLE
 
         statWrapper.visibility = if (usesTaskAttributeStats) View.VISIBLE else View.GONE
         if (isCreating) {
@@ -232,6 +269,10 @@ class TaskFormActivity : BaseActivity() {
     }
 
     private fun fillForm(task: Task) {
+        if (!task.isValid) {
+            return
+        }
+        canSave = true
         textEditText.setText(task.text)
         notesEditText.setText(task.notes)
         taskDifficultyButtons.selectedDifficulty = task.priority
@@ -239,7 +280,11 @@ class TaskFormActivity : BaseActivity() {
             Task.TYPE_HABIT -> {
                 habitScoringButtons.isPositive = task.up ?: false
                 habitScoringButtons.isNegative = task.down ?: false
-                task.frequency?.let { habitResetStreakButtons.selectedResetOption = HabitResetOption.valueOf(it.toUpperCase()) }
+                task.frequency?.let {
+                    if (it.isNotBlank()) {
+                        habitResetStreakButtons.selectedResetOption = HabitResetOption.valueOf(it.toUpperCase(Locale.US))
+                    }
+                }
                 habitAdjustPositiveStreakView.setText((task.counterUp ?: 0).toString())
                 habitAdjustNegativeStreakView.setText((task.counterDown ?: 0).toString())
             }
@@ -291,10 +336,15 @@ class TaskFormActivity : BaseActivity() {
     }
 
     private fun saveTask() {
+        if (isSaving) {
+            return
+        }
+        isSaving = true
         var thisTask = task
         if (thisTask == null) {
             thisTask = Task()
             thisTask.type = taskType
+            thisTask.dateCreated = Date()
         }
         thisTask.text = textEditText.text.toString()
         thisTask.notes = notesEditText.text.toString()
@@ -319,31 +369,51 @@ class TaskFormActivity : BaseActivity() {
             thisTask.value = rewardValueFormView.value
         }
 
-        if (taskType == Task.TYPE_DAILY || taskType == Task.TYPE_TODO) {
-            thisTask.checklist = checklistContainer.checklistItems
-            thisTask.reminders = remindersContainer.reminders
-        }
-        thisTask.tags = RealmList()
-        tagsWrapper.forEachIndexed { index, view ->
-            val tagView = view as? CheckBox
-            if (tagView?.isChecked == true) {
-                thisTask.tags?.add(tags[index])
+        val resultIntent = Intent()
+        resultIntent.putExtra(TASK_TYPE_KEY, taskType)
+        if (!isChallengeTask) {
+            if (taskType == Task.TYPE_DAILY || taskType == Task.TYPE_TODO) {
+                thisTask.checklist = checklistContainer.checklistItems
+                thisTask.reminders = remindersContainer.reminders
             }
+            thisTask.tags = RealmList()
+            tagsWrapper.forEachIndexed { index, view ->
+                val tagView = view as? CheckBox
+                if (tagView?.isChecked == true) {
+                    thisTask.tags?.add(tags[index])
+                }
+            }
+
+            if (isCreating) {
+                taskRepository.createTaskInBackground(thisTask)
+            } else {
+                taskRepository.updateTaskInBackground(thisTask)
+            }
+
+            if (thisTask.type == Task.TYPE_DAILY || thisTask.type == Task.TYPE_TODO) {
+                taskAlarmManager.scheduleAlarmsForTask(thisTask)
+            }
+        } else {
+                resultIntent.putExtra(PARCELABLE_TASK, thisTask)
         }
 
-        if (isCreating) {
-            taskRepository.createTaskInBackground(thisTask)
-        } else {
-            taskRepository.updateTaskInBackground(thisTask)
-        }
-        dismissKeyboard()
-        finish()
+        val mainHandler = Handler(this.mainLooper)
+        mainHandler.postDelayed({
+            setResult(Activity.RESULT_OK, resultIntent)
+            dismissKeyboard()
+            finish()
+        }, 500)
     }
 
     private fun deleteTask() {
-        task?.id?.let { taskRepository.deleteTask(it).subscribe(Consumer {  }, RxErrorHandler.handleEmptyError()) }
-        dismissKeyboard()
-        finish()
+        AlertDialog.Builder(this)
+                .setTitle(R.string.are_you_sure)
+                .setPositiveButton(R.string.delete_task) { _, _ ->
+                    task?.id?.let { taskRepository.deleteTask(it).subscribe(Consumer {  }, RxErrorHandler.handleEmptyError()) }
+                    dismissKeyboard()
+                    finish()
+                }
+                .setNegativeButton(android.R.string.cancel, null).show()
     }
 
     private fun dismissKeyboard() {
@@ -356,8 +426,11 @@ class TaskFormActivity : BaseActivity() {
 
     companion object {
         const val TASK_ID_KEY = "taskId"
+        const val USER_ID_KEY = "userId"
         const val TASK_TYPE_KEY = "type"
-        const val SHOW_TAG_SELECTION = "show_tag_selection"
+        const val IS_CHALLENGE_TASK = "isChallengeTask"
+
+        const val PARCELABLE_TASK = "parcelable_task"
 
         // in order to disable the event handler in MainActivity
         const val SET_IGNORE_FLAG = "ignoreFlag"
